+++ Framework-Headers/PrivateFrameworks/iWorkImport._framework/Objc/TNPageLayout.h
@@ -45,7 +45,7 @@ __attribute__((visibility("hidden")))
 + (id)pageLayoutWithPageController:(id)arg1 pageCoordinate:(struct TSUCellCoord)arg2;
 - (void)addChild:(id)arg1;
 - (struct CGSize)adjustedInsetsForTarget:(id)arg1;
-- (unsigned int)autosizeFlagsForTextLayout:(id)arg1;
+- (unsigned long long)autosizeFlagsForTextLayout:(id)arg1;
 - (struct CGRect)autosizedFrameForTextLayout:(id)arg1 textSize:(struct CGSize)arg2;
 - (struct CGRect)bodyRect;
 - (id)computeLayoutGeometry;
