+++ Framework-Headers/PrivateFrameworks/iWorkImport._framework/Objc/TSWPStorageAttributeRangeProvider.h
@@ -13,7 +13,7 @@ __attribute__((visibility("hidden")))
 }
 
 - (void)dealloc;
-- (id)initWithStorage:(id)arg1 kind:(int)arg2;
+- (id)initWithStorage:(id)arg1 kind:(unsigned long long)arg2;
 - (void)nextRange;
 
 @end
